refactor: rename flutter_inspector_impl to flutter_inspector
Rename the core implementation file (and its test) to drop the redundant _impl
suffix, since the class it defines is already named FlutterInspector. Update
all export/import references across lib and test accordingly.
