+++ lib/flutter_inspector_kit.dart
@@ -1,4 +1,4 @@
-export 'src/core/flutter_inspector_impl.dart';
+export 'src/core/flutter_inspector.dart';
 export 'src/interceptors/dio_interceptor.dart';
 export 'src/models/database_browser_source.dart';
 export 'src/models/database_operation.dart';
