+++ lib/src/core/flutter_inspector.dart
